feat: improved symlink support
- Every symlink found in a node_modules directory (or a scope thereof) is treated as a project root
- All symlink paths (as opposed to only node_modules) are now cached in `data.links`
- Simplify and add more comments
- Find and watch linked dependencies before crawling them all in parallel

Tested manually on a project using PNPM and yarn link.

"Fresh install" startup perf could be improved by searching for linked dependencies in parallel instead of in batches.

When `watchman watch-del-all` is not used, startup perf is lovely. The difference between using and not using the `--reset-cache` flag is negligible. Startup perf for projects _not_ using symlinks remains unaffected.